+++ cookbooks/videolan_vlc/metadata.rb
@@ -1,6 +1,7 @@
+name			 "videolan_vlc"
 maintainer       "Wesleyan University"
 maintainer_email "software@wesleyan.edu"
 license          "All rights reserved"
 description      "Installs VLC and registers file associations"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.5.0"
+version          "0.6.0"