+++ cookbooks/videolan_vlc/recipes/win.rb
@@ -1,6 +1,13 @@
-# Windows installer for VLC
-
+#
+# Cookbook Name:: videon_vlc
+# Recipe:: win
+#
+# Copyright 2013, Wesleyan University
+#
+# All rights reserved - Do Not Redistribute
+#
+# Install package
 windows_package "VideoLAN VLC" do
-  source "http://ims-chef.wesleyan.edu/windows/videolan_vlc/vlc-2.0.6-win32.exe"
-  version "2.0.6"
+  source "http://ims-chef.wesleyan.edu/windows/videolan_vlc/vlc-2.0.7-win32.exe"
+  version "2.0.7"
 end